requirements.txt
click==8.0.4
Flask==2.0.3
Flask-PyMongo==2.3.0
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
pymongo==4.0.2
python-dotenv==0.20.0
Werkzeug==2.0.3
